web3py-favorites-cu/pyproject.toml
[project]
name = "web3py-favorites-cu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "titanoboa>=0.2.5",
    "vyper>=0.4.0",
    "web3>=7.7.0",
]
